More PR feedback:
- A few comments/whitespace/formatting issues.
- Share code between PutAsync()/PostAsync().
- Move QueryDefinition extensions that are specifically for datasets to the `Keen.Core.Dataset` namespace.
