+++ Keen/Keen.csproj
@@ -57,6 +57,7 @@
     <Compile Include="Dataset\Datasets.cs" />
     <Compile Include="DataEnrichment\EventAddOn.cs" />
     <Compile Include="Dataset\IDataset.cs" />
+    <Compile Include="Dataset\QueryDefinitionExtensions.cs" />
     <Compile Include="HttpClientCache.cs" />
     <Compile Include="IHttpClientProvider.cs" />
     <Compile Include="IKeenHttpClient.cs" />
