+++ Keen/KeenClient.cs
@@ -1033,7 +1033,7 @@ namespace Keen.Core
         }
 
         /// <summary>
-        /// Lists the first n cached dataset definitions in your project
+        /// Lists the first n cached dataset definitions in your project.
         /// </summary>
         /// <param name="limit">How many cached dataset definitions to return at a time (1-100). Defaults to 10.</param>
         /// <param name="afterName">A cursor for use in pagination. afterName is the Cached Dataset name that defines your place in the list. 
@@ -1048,7 +1048,7 @@ namespace Keen.Core
         }
 
         /// <summary>
-        /// Lists the first n cached dataset definitions in your project
+        /// Lists the first n cached dataset definitions in your project.
         /// </summary>
         /// <param name="limit">How many cached dataset definitions to return at a time (1-100). Defaults to 10.</param>
         /// <param name="afterName">A cursor for use in pagination. afterName is the Cached Dataset name that defines your place in the list.
