+++ Keen/Query/QueryDefinition.cs
@@ -52,29 +52,5 @@ namespace Keen.Core.Query
         public IEnumerable<string> GroupBy { get; set; }
     }
 
-    internal static class QueryDefinitionExtensions
-    {
-        public static void ValidateForCachedDataset(this QueryDefinition query)
-        {
-            if (string.IsNullOrWhiteSpace(query.AnalysisType))
-            {
-                throw new KeenException("QueryDefinition must have an analysis type");
-            }
-
-            if (string.IsNullOrWhiteSpace(query.EventCollection))
-            {
-                throw new KeenException("QueryDefinition must specify an event collection");
-            }
-
-            if (string.IsNullOrWhiteSpace(query.Timeframe))
-            {
-                throw new KeenException("QueryDefinition must specify a timeframe");
-            }
-
-            if (string.IsNullOrWhiteSpace(query.Interval))
-            {
-                throw new KeenException("QueryDefinition must specify an interval");
-            }
-        }
-    }
+    
 }
